Forced script shebang rewriter to point to /usr/bin/python
Angstrom does not have a python2 symlink (only python and python2.6).  Since python on Arch points to python3, attempting to build through python2 caused scripts to get python2 in the shebang line; this made them impossible to execute on Pandora's Angstrom.  If Angstrom included a python2 symlink, this change could be removed.