Removed runtime specification of "command-packages" in the readme.
Turns out command-packages must be specified in the setup.cfg in order for
bdist_pnd to be able to call gen_pxml.  Specifying command-packages on the
command line makes the gen_pxml call fail silently, causing the whole thing to
fail shortly thereafter when PXML.xml can't be found.  So I won't recommend
using the command line for command-packages.