Updated gen_pxml to match latest changes to PXML spec.
The most notable change is rearranging PXML output to put elements in the
recommended order.  This isn't necessary for a PND to run, but it allows the
PXML to be successfully validated against the schema file.  I'll hopefully add
that as a test one day.  All code (including the options list and finalize
method) were put in to this same order.

The major change to the spec is the allowance of non-numeric characters in the
version number.  I wasn't previously checking for this (as is obvious from my
test cases), but now I make sure to error on bad characters.

I'm not sure if it was always this way, but having an author name is now
mandatory.  Error is raised if it's missing.

Finally, the mkdir element is no longer in the spec, so it's been removed from
here.  I never liked that one anyways.